Refactor samples.go and code cleaning.
