+++ pkg/pgclient/client.go
@@ -140,12 +140,13 @@ func NewClientWithPool(cfg *Config, numCopiers int, connPool *pgxpool.Pool, mt t
 	var dbIngestor *ingestor.DBIngestor
 	if !readOnly {
 		var (
-			err                 error
+			err error
+
 			exemplarKeyPosCache = cache.NewExemplarLabelsPosCache(cfg.CacheConfig)
 		)
 		dbIngestor, err = ingestor.NewPgxIngestor(dbConn, metricsCache, seriesCache, exemplarKeyPosCache, &c)
 		if err != nil {
-			log.Error("msg", "err starting ingestor", "err", err)
+			log.Error("msg", "err starting the ingestor", "err", err)
 			return nil, err
 		}
 	}
