+++ pkg/pgmodel/querier/querier.go
@@ -85,7 +85,7 @@ type pgxQuerier struct {
 
 var _ Querier = (*pgxQuerier)(nil)
 
-// Select implements the Querier interface. It is the entry point for our
+// SelectSamples implements the Querier interface. It is the entry point for our
 // own version of the Prometheus engine.
 func (q *pgxQuerier) Select(mint int64, maxt int64, sortSeries bool, hints *storage.SelectHints, qh *QueryHints, path []parser.Node, ms ...*labels.Matcher) (SeriesSet, parser.Node) {
 	rows, topNode, err := q.getResultRows(mint, maxt, hints, qh, path, ms)
@@ -96,6 +96,14 @@ func (q *pgxQuerier) Select(mint int64, maxt int64, sortSeries bool, hints *stor
 	ss := buildSeriesSet(rows, q.labelsReader)
 	return ss, topNode
 }
+
+//func (q *pgxQuerier) SelectExemplars(mint, max int64, ms ...*labels.Matcher) error {
+//	metrics, seriesIDs, err := GetMetricNameSeriesIDFromMatchers(q.conn, ms)
+//	if err != nil {
+//		return fmt.Errorf("get metrics-seriesIDs from matchers: %w", err)
+//	}
+//
+//}
 
 // Query implements the Querier interface. It is the entry point for
 // remote-storage queries.
